feat(social): add user profile components and refactor layout
Introduce new components for user profile, including ProfileSection, RecentActivities, Saved, Spaces, TeamSelector, UserBadges, and UserTier. Refactor the social layout to utilize these components and integrate user data fetching via a new `getUserInfo` hook. Update global styles for improved background color consistency.
