+++ src/app/(social)/_components/ProfileSection.tsx
@@ -1,0 +1,40 @@
+import { Star } from 'lucide-react';
+import React from 'react';
+import Image from 'next/image';
+import { User } from '@/lib/api/models/user';
+import TeamSelector from './TeamSelector';
+import UserTier from './UserTier';
+import UserBadges from './UserBadges';
+
+interface ProfileSectionProps {
+  user: User;
+}
+
+export default function ProfileSection({ user }: ProfileSectionProps) {
+  return (
+    <div className="flex flex-col gap-5 px-4 py-5 rounded-[10px] bg-component-bg">
+      <TeamSelector />
+
+      <div className="relative">
+        <Image
+          src={user.profile_url || '/default-profile.png'}
+          alt={user.nickname}
+          width={80}
+          height={80}
+          className="rounded-full border-2 border-primary object-cover object-top"
+        />
+      </div>
+
+      <div className="font-medium">{user?.nickname}</div>
+
+      <div
+        id="user-profile-description"
+        className="text-xs text-gray-400"
+        dangerouslySetInnerHTML={{ __html: user.html_contents }}
+      />
+
+      <UserTier />
+      <UserBadges />
+    </div>
+  );
+}
