+++ src/assets/css/globals.css
@@ -17,7 +17,7 @@ a {
 
 @theme static {
   /* Tailwindcss variables */
-  --color-background: #1a1a1a;
+  --color-background: #1e1e1e;
   --color-primary: #fcb300;
   --color-secondary: #777677;
 
@@ -40,7 +40,7 @@ a {
   --color-c-p-50-10: #db27801a;
   --color-supportive: #6eedd8;
   --color-against: #ff9eee;
-  --color-bg: #1a1a1a;
+  --color-bg: #1e1e1e;
   --color-tab-hover: #282828;
   --color-mission-title: #dee1e6;
   --color-tag: #dee1e6;
